Moved project to Visual Studio 2015

+++ intern/src/core/core_vector3.cpp
@@ -180,7 +180,7 @@ namespace Core
 
     float CVector3::GetLength()
 	{
-        return sqrt(GetDotProduct(*this));
+        return sqrtf(GetDotProduct(*this));
 	}
 
     CVector3& CVector3::Normalize()
